fix remove template lifecycle
